FEAT (jhonatan): Refactorizar delet
